+++ UnityProject/Assets/Scenes/game.unity
@@ -373,6 +373,7 @@ GameObject:
   - component: {fileID: 673763353}
   - component: {fileID: 673763352}
   - component: {fileID: 673763351}
+  - component: {fileID: 673763358}
   m_Layer: 0
   m_Name: Canvas
   m_TagString: Untagged
@@ -520,6 +521,18 @@ RectTransform:
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0, y: 0}
+--- !u!114 &673763358
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 673763350}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 6d3cb24e578bd654db5ddddaf941542d, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  FindCameraName: Main Camera
 --- !u!1 &836716736
 GameObject:
   m_ObjectHideFlags: 0
