+++ UnityProject/Assets/Script/NoteSetting.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 6d3cb24e578bd654db5ddddaf941542d
+timeCreated: 1508994430
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
